requirements.txt
streamlit==1.28.0
groq==0.4.1
PyPDF2==3.0.1
python-dotenv==1.0.0
pandas==2.1.0
python-docx==0.8.11
langchain==0.0.350
tiktoken==0.5.1
requests==2.31.0
openai == 1.3.0